requirements.txt
python-telegram-bot==20.3
python-docx==0.8.11